move getTopicPolicies method to BrokerService.java
